requirements.txt
Flask==3.0.0
flask-cors==4.0.0

# Google Sheets Integration (optional - for lead data export)
gspread==5.12.0
google-auth>=2.23.4
python-dotenv==1.0.0

